fix: turn getProof task into module
